+++ package.json
@@ -38,7 +38,7 @@
     "lucide-react": "^0.453.0",
     "negotiator": "^0.6.3",
     "next": "^14.2.4",
-    "next-auth": "^4.24.7",
+    "next-auth": "^5.0.0-beta.25",
     "next-themes": "^0.3.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
