--- src/server/auth.ts
@@ -1,134 +1,0 @@
-import { DrizzleAdapter } from "@auth/drizzle-adapter";
-import {
-  getServerSession,
-  // User,
-  type DefaultSession,
-  type NextAuthOptions,
-} from "next-auth";
-import { type Adapter } from "next-auth/adapters";
-import Credentials from "next-auth/providers/credentials"
-
-// import { env } from "@/env";
-import { db } from "@/server/db";
-import { accounts } from "./db/schema/accounts";
-import { sessions } from "./db/schema/sessions";
-import { users } from "./db/schema/users";
-import { verificationTokens } from "./db/schema/verificationTokens";
-
-type UserRole = "USER" | "WORKER" | "ADMIN";
-
-/**
- * Module augmentation for `next-auth` types. Allows us to add custom properties to the `session`
- * object and keep type safety.
- *
- * @see https://next-auth.js.org/getting-started/typescript#module-augmentation
- */
-declare module "next-auth" {
-  interface Session extends DefaultSession {
-    user: {
-      id: string;
-      // ...other properties
-      role: UserRole;
-    } & DefaultSession["user"];
-  }
-
-  interface User {
-    // ...other properties
-    role: UserRole;
-  }
-}
-
-/**
- * Options for NextAuth.js used to configure adapters, providers, callbacks, etc.
- *
- * @see https://next-auth.js.org/configuration/options
- */
-export const authOptions: NextAuthOptions = {
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = user.id;
-        token.email = user.email;
-        token.name = user.name;
-        token.role = user.role;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (token && session.user) {
-        session.user.id = token.id as string;
-        session.user.email = token.email!;
-        session.user.name = token.name!;
-        session.user.role = token.role as UserRole;
-      }
-      return session;
-    },
-  },
-  session: {
-    strategy: "jwt",
-    maxAge: 30 * 24 * 60 * 60, // 30 days
-  },
-  adapter: DrizzleAdapter(db, {
-    usersTable: users,
-    accountsTable: accounts,
-    sessionsTable: sessions,
-    verificationTokensTable: verificationTokens,
-  }) as Adapter,
-  providers: [
-    Credentials({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "text" },
-        password: { label: "Password", type: "password" },
-      },
-      authorize: async (credentials) => {
-        console.log("logging in ...");
-        if (credentials?.email === "admin" && credentials?.password === "admin") {
-          return {
-            id: "1",
-            name: "Admin",
-            email: "admin@example.com",
-            role: "ADMIN" as UserRole,
-          };
-        }
-        if (credentials?.email === "worker" && credentials?.password === "worker") {
-          return {
-            id: "2",
-            name: "Worker",
-            email: "worker@example.com",
-            role: "WORKER" as UserRole,
-          };
-        }
-        if (credentials?.email === "user" && credentials?.password === "user") {
-          return {
-            id: "3",
-            name: "User",
-            email: "user@example.com",
-            role: "USER" as UserRole,
-          };
-        }
-        return null;
-      },
-    }),
-    // DiscordProvider({
-    //   clientId: env.DISCORD_CLIENT_ID,
-    //   clientSecret: env.DISCORD_CLIENT_SECRET,
-    // }),
-    /**
-     * ...add more providers here.
-     *
-     * Most other providers require a bit more work than the Discord provider. For example, the
-     * GitHub provider requires you to add the `refresh_token_expires_in` field to the Account
-     * model. Refer to the NextAuth.js docs for the provider you want to use. Example:
-     *
-     * @see https://next-auth.js.org/providers/github
-     */
-  ],
-};
-
-/**
- * Wrapper for `getServerSession` so that you don't need to import the `authOptions` in every file.
- *
- * @see https://next-auth.js.org/configuration/nextjs
- */
-export const getServerAuthSession = () => getServerSession(authOptions);
